Secure-Document-Platform/document-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-jose
requests
python-multipart